darkmode without favorites
